Module Peer_tube.Client

val get_abuses : ?id:string -> ?predefined_reason:string -> ?search:string -> ?state:string -> ?search_reporter:string -> ?search_reportee:string -> ?search_video:string -> ?search_video_channel:string -> ?video_is:string -> ?filter:string -> ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List abuses

parameter id
only list the report with this id

parameter predefined_reason
predefined reason the listed reports should contain

parameter search
plain search that will match with video titles, reporter names and more

parameter search_reporter
only list reports of a specific reporter

parameter search_reportee
only list reports of a specific reportee

parameter search_video
only list reports of a specific video

parameter search_video_channel
only list reports of a specific video channel

parameter video_is
only list deleted or blocklisted videos

parameter filter
only list account, comment or video reports

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort abuses by criteria

val post_api_v1_abuses : t -> unit -> Jsont.json

Report an abuse

val put_api_v1_abuses : abuse_id:string -> t -> unit -> Jsont.json

Update an abuse

parameter abuse_id
Abuse id

val delete_api_v1_abuses : abuse_id:string -> t -> unit -> Jsont.json

Delete an abuse

parameter abuse_id
Abuse id

val get_api_v1_abuses_messages : abuse_id:string -> t -> unit -> Jsont.json

List messages of an abuse

parameter abuse_id
Abuse id

val post_api_v1_abuses_messages : abuse_id:string -> t -> unit -> Jsont.json

Add message to an abuse

parameter abuse_id
Abuse id

val delete_api_v1_abuses_messages : abuse_id:string -> abuse_message_id:string -> t -> unit -> Jsont.json

Delete an abuse message

parameter abuse_id
Abuse id

parameter abuse_message_id
Abuse message id

val get_accounts : ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List accounts

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val get_account_followers : name:string -> ?start:string -> ?count:string -> ?sort:string -> ?search:string -> t -> unit -> Jsont.json

List followers of an account

parameter name
The username or handle of the account

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort followers by criteria

parameter search
Plain text search, applied to various parts of the model depending on endpoint

val get_api_v1_accounts_video_playlists : name:string -> ?start:string -> ?count:string -> ?sort:string -> ?search:string -> ?playlist_type:string -> ?include_collaborations:string -> ?channel_name_one_of:string -> t -> unit -> Jsont.json

List playlists of an account

parameter name
The username or handle of the account

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

parameter search
Plain text search, applied to various parts of the model depending on endpoint

parameter include_collaborations
**PeerTube >= 8.0** Include objects from collaborated channels

parameter channel_name_one_of
**PeerTube >= 8.0** Filter on playlists that are published on a channel with one of these names

val put_api_v1_automatic_tags_policies_accounts_comments : account_name:string -> t -> unit -> Jsont.json

Update account auto tag policies on comments

**PeerTube >= 6.2**

parameter account_name
account name to update auto tag policies

val update_client_language : t -> unit -> Jsont.json

Update client language

Set a cookie so that, the next time the client refreshes the HTML of the web interface, PeerTube will use the next language

val put_custom_config : t -> unit -> Jsont.json

Set instance runtime configuration

val del_custom_config : t -> unit -> Jsont.json

Delete instance runtime configuration

val delete_api_v1_config_instance_avatar : t -> unit -> Jsont.json

Delete instance avatar

val post_api_v1_config_instance_avatar_pick : t -> unit -> Jsont.json

Update instance avatar

val delete_api_v1_config_instance_banner : t -> unit -> Jsont.json

Delete instance banner

val post_api_v1_config_instance_banner_pick : t -> unit -> Jsont.json

Update instance banner

val delete_api_v1_config_instance_logo_logo_type : logo_type:string -> t -> unit -> Jsont.json

Delete instance logo

val post_api_v1_config_instance_logo_logo_type_pick : logo_type:string -> t -> unit -> Jsont.json

Update instance logo

val put_api_v1_custom_pages_homepage_instance : t -> unit -> Jsont.json

Set instance custom homepage

val post_api_v1_jobs_pause : t -> unit -> Jsont.json

Pause job queue

val post_api_v1_jobs_resume : t -> unit -> Jsont.json

Resume job queue

val get_jobs : state:string -> ?job_type:string -> ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List instance jobs

parameter state
The state of the job ('' for for no filter)

parameter job_type
job type

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val post_api_v1_metrics_playback : body:PlaybackMetric.Create.t -> t -> unit -> Jsont.json

Create playback metrics

These metrics are exposed by OpenTelemetry metrics exporter if enabled.

val add_plugin : t -> unit -> Jsont.json

Install a plugin

val uninstall_plugin : t -> unit -> Jsont.json

Uninstall a plugin

val update_plugin : t -> unit -> Jsont.json

Update a plugin

val get_api_v1_plugins_public_settings : npm_name:string -> t -> unit -> Jsont.json

Get a plugin's public settings

parameter npm_name
name of the plugin/theme on npmjs.com or in its package.json

val get_api_v1_plugins_registered_settings : npm_name:string -> t -> unit -> Jsont.json

Get a plugin's registered settings

parameter npm_name
name of the plugin/theme on npmjs.com or in its package.json

val put_api_v1_plugins_settings : npm_name:string -> t -> unit -> Jsont.json

Set a plugin's settings

parameter npm_name
name of the plugin/theme on npmjs.com or in its package.json

val get_api_v1_runners : ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List runners

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort runners by criteria

val get_api_v1_runners_jobs : ?start:string -> ?count:string -> ?sort:string -> ?search:string -> ?state_one_of:string -> t -> unit -> Jsont.json

List jobs

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort runner jobs by criteria

parameter search
Plain text search, applied to various parts of the model depending on endpoint

val post_api_v1_runners_jobs_request : t -> unit -> Jsont.json

Request a new job

API used by PeerTube runners

val delete_api_v1_runners_jobs : job_uuid:string -> t -> unit -> Jsont.json

Delete a job

The endpoint will first cancel the job if needed, and then remove it from the database. Children jobs will also be removed

val post_api_v1_runners_jobs_abort : job_uuid:string -> t -> unit -> Jsont.json

Abort job

API used by PeerTube runners

val post_api_v1_runners_jobs_accept : job_uuid:string -> t -> unit -> Jsont.json

Accept job

API used by PeerTube runners

val get_api_v1_runners_jobs_cancel : job_uuid:string -> t -> unit -> Jsont.json

Cancel a job

val post_api_v1_runners_jobs_error : job_uuid:string -> t -> unit -> Jsont.json

Post job error

API used by PeerTube runners

val post_api_v1_runners_jobs_success : job_uuid:string -> t -> unit -> Jsont.json

Post job success

API used by PeerTube runners

val post_api_v1_runners_jobs_update : job_uuid:string -> t -> unit -> Jsont.json

Update job

API used by PeerTube runners

val post_api_v1_runners_register : t -> unit -> Jsont.json

Register a new runner

API used by PeerTube runners

val get_api_v1_runners_registration_tokens : ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List registration tokens

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort registration tokens by criteria

val post_api_v1_runners_registration_tokens_generate : t -> unit -> Jsont.json

Generate registration token

Generate a new runner registration token

val delete_api_v1_runners_registration_tokens : registration_token_id:string -> t -> unit -> Jsont.json

Remove registration token

Remove a registration token. Runners that used this token for their registration are automatically removed.

val post_api_v1_runners_unregister : t -> unit -> Jsont.json

Unregister a runner

API used by PeerTube runners

val delete_api_v1_runners : runner_id:string -> t -> unit -> Jsont.json

Delete a runner

val search_playlists : search:string -> ?start:string -> ?count:string -> ?search_target:string -> ?sort:string -> ?host:string -> ?uuids:string -> t -> unit -> Jsont.json

Search playlists

parameter search
String to search. If the user can make a remote URI search, and the string is an URI then the PeerTube instance will fetch the remote object and add it to its database. Then, you can use the REST API to fetch the complete playlist information and interact with it.

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter search_target
If the administrator enabled search index support, you can override the default search target.

**Warning**: If you choose to make an index search, PeerTube will get results from a third party service. It means the instance may not yet know the objects you fetched. If you want to load video/channel information: * If the current user has the ability to make a remote URI search (this information is available in the config endpoint), then reuse the search API to make a search using the object URI so PeerTube instance fetches the remote object and fill its database. After that, you can use the classic REST API endpoints to fetch the complete object or interact with it * If the current user doesn't have the ability to make a remote URI search, then redirect the user on the origin instance or fetch the data from the origin instance API

parameter sort
Sort column

parameter host
Find elements owned by this host

parameter uuids
Find elements with specific UUIDs

val get_instance_audit_logs : t -> unit -> Jsont.json

Get instance audit logs

val get_api_v1_server_blocklist_accounts : ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List account blocks

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val post_api_v1_server_blocklist_accounts : t -> unit -> Jsont.json

Block an account

val delete_api_v1_server_blocklist_accounts : account_name:string -> t -> unit -> Jsont.json

Unblock an account by its handle

parameter account_name
account to unblock, in the form `username@domain`

val get_api_v1_server_blocklist_servers : ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List server blocks

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val post_api_v1_server_blocklist_servers : t -> unit -> Jsont.json

Block a server

val delete_api_v1_server_blocklist_servers : host:string -> t -> unit -> Jsont.json

Unblock a server by its domain

parameter host
server domain to unblock

val get_api_v1_server_followers : ?state:string -> ?actor_type:string -> ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List instances following the server

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val delete_api_v1_server_followers : handle:string -> t -> unit -> Jsont.json

Remove or reject a follower to your server

parameter handle
The remote actor handle to remove from your followers

val post_api_v1_server_followers_accept : handle:string -> t -> unit -> Jsont.json

Accept a pending follower to your server

parameter handle
The remote actor handle to remove from your followers

val post_api_v1_server_followers_reject : handle:string -> t -> unit -> Jsont.json

Reject a pending follower to your server

parameter handle
The remote actor handle to remove from your followers

val get_api_v1_server_following : ?state:string -> ?actor_type:string -> ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List instances followed by the server

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val post_api_v1_server_following : t -> unit -> Jsont.json

Follow a list of actors (PeerTube instance, channel or account)

val delete_api_v1_server_following : host_or_handle:string -> t -> unit -> Jsont.json

Unfollow an actor (PeerTube instance, channel or account)

parameter host_or_handle
The hostOrHandle to unfollow

val get_instance_logs : t -> unit -> Jsont.json

Get instance logs

val send_client_log : body:SendClientLog.T.t -> t -> unit -> Jsont.json

Send client log

val put_mirrored_video : t -> unit -> Jsont.json

Mirror a video

val del_mirrored_video : redundancy_id:string -> t -> unit -> Jsont.json

Delete a mirror done on a video

parameter redundancy_id
id of an existing redundancy on a video

val put_api_v1_server_redundancy : host:string -> t -> unit -> Jsont.json

Update a server redundancy policy

parameter host
server domain to mirror

val post_api_v1_users_ask_reset_password : t -> unit -> Jsont.json

Ask to reset password

An email containing a reset password link

val resend_email_to_verify_user : t -> unit -> Jsont.json

Resend user verification link

val put_user_info : body:UpdateMe.T.t -> t -> unit -> Jsont.json

Update my user information

val get_my_abuses : ?id:string -> ?state:string -> ?sort:string -> ?start:string -> ?count:string -> t -> unit -> Jsont.json

List my abuses

parameter id
only list the report with this id

parameter sort
Sort abuses by criteria

parameter start
Offset used to paginate results

parameter count
Number of items to return

val delete_api_v1_users_me_avatar : t -> unit -> Jsont.json

Delete my avatar

val post_api_v1_users_me_avatar_pick : t -> unit -> Jsont.json

Update my user avatar

val post_api_v1_users_me_history_videos_remove : t -> unit -> Jsont.json

Clear video history

val delete_api_v1_users_me_history_videos : video_id:string -> t -> unit -> Jsont.json

Delete history element

val post_api_v1_users_me_new_feature_info_read : t -> unit -> Jsont.json

Mark feature info as read

**PeerTube >= v8.0.0

val put_api_v1_users_me_notification_settings : body:UserNotificationSettings.T.t -> t -> unit -> Jsont.json

Update my notification settings

val post_api_v1_users_me_notifications_read : t -> unit -> Jsont.json

Mark notifications as read by their id

val post_api_v1_users_me_notifications_read_all : t -> unit -> Jsont.json

Mark all my notification as read

val post_api_v1_users_me_subscriptions : t -> unit -> Jsont.json

Add subscription to my user

val get_api_v1_users_me_subscriptions_exist : uris:string -> t -> unit -> Jsont.json

Get if subscriptions exist for my user

parameter uris
list of uris to check if each is part of the user subscriptions

val delete_api_v1_users_me_subscriptions : subscription_handle:string -> t -> unit -> Jsont.json

Delete subscription of my user

parameter subscription_handle
The subscription handle

val get_api_v1_users_me_video_playlists_videos_exist : video_ids:string -> t -> unit -> Jsont.json

Check video exists in my playlists

parameter video_ids
The video ids to check

val get_api_v1_users_me_video_quota_used : t -> unit -> Jsont.json

Get my user used quota

val get_api_v1_users_me_videos_comments : ?search:string -> ?search_account:string -> ?search_video:string -> ?video_id:string -> ?video_channel_id:string -> ?auto_tag_one_of:string -> ?is_held_for_review:string -> ?include_collaborations:string -> t -> unit -> Jsont.json

List comments on user's videos

**PeerTube >= 6.2**

parameter search
Plain text search, applied to various parts of the model depending on endpoint

parameter search_account
Filter comments by searching on the account

parameter search_video
Filter comments by searching on the video

parameter video_id
Limit results on this specific video

parameter video_channel_id
Limit results on this specific video channel

parameter auto_tag_one_of
**PeerTube >= 6.2** filter on comments that contain one of these automatic tags

parameter is_held_for_review
only display comments that are held for review

parameter include_collaborations
**PeerTube >= 8.0** Include objects from collaborated channels

val register_user : body:RegisterUser.T.t -> t -> unit -> Jsont.json

Register a user

Signup has to be enabled and signup approval is not required

val list_registrations : ?start:string -> ?count:string -> ?search:string -> ?sort:string -> t -> unit -> Jsont.json

List registrations

parameter start
Offset used to paginate results

parameter count
Number of items to return

val resend_email_to_verify_registration : t -> unit -> Jsont.json

Resend verification link to registration request email

val delete_registration : registration_id:string -> t -> unit -> Jsont.json

Delete registration

Delete the registration entry. It will not remove the user associated with this registration (if any)

parameter registration_id
Registration ID

val accept_registration : registration_id:string -> body:UserRegistrationAcceptOrReject.T.t -> t -> unit -> Jsont.json

Accept registration

parameter registration_id
Registration ID

val reject_registration : registration_id:string -> body:UserRegistrationAcceptOrReject.T.t -> t -> unit -> Jsont.json

Reject registration

parameter registration_id
Registration ID

val verify_registration_email : registration_id:string -> t -> unit -> Jsont.json

Verify a registration email

Following a user registration request, the user will receive an email asking to click a link containing a secret.

parameter registration_id
Registration ID

val revoke_oauth_token : t -> unit -> Jsont.json

Logout

Revokes your access token and its associated refresh token, destroying your current session.

val get_oauth_token : t -> unit -> Jsont.json

Login

With your client id and secret(#operation/getOAuthClient), you can retrieve an access and refresh tokens.

val get_user : id:string -> ?with_stats:string -> t -> unit -> Jsont.json

Get a user

parameter id
Entity id

parameter with_stats
include statistics about the user (only available as a moderator/admin)

val put_user : id:string -> body:UpdateUser.T.t -> t -> unit -> Jsont.json

Update a user

parameter id
Entity id

val del_user : id:string -> t -> unit -> Jsont.json

Delete a user

parameter id
Entity id

val post_api_v1_users_reset_password : id:string -> t -> unit -> Jsont.json

Reset password

parameter id
Entity id

val get_api_v1_users_token_sessions : id:string -> t -> unit -> Jsont.json

List token sessions

parameter id
Entity id

val get_api_v1_users_token_sessions_revoke : id:string -> token_session_id:string -> t -> unit -> Jsont.json

List token sessions

parameter id
Entity id

parameter token_session_id
Token session Id

val confirm_two_factor_request : id:string -> t -> unit -> Jsont.json

Confirm two factor auth

Confirm a two factor authentication request

parameter id
Entity id

val disable_two_factor : id:string -> t -> unit -> Jsont.json

Disable two factor auth

Disable two factor authentication of a user

parameter id
Entity id

val verify_user : id:string -> t -> unit -> Jsont.json

Verify a user

Following a user registration, the new user will receive an email asking to click a link containing a secret. This endpoint can also be used to verify a new email set in the user account.

parameter id
Entity id

val list_user_exports : user_id:string -> t -> unit -> Jsont.json

List user exports

**PeerTube >= 6.1**

parameter user_id
User id

val request_user_export : user_id:string -> t -> unit -> Jsont.json

Request user export

Request an archive of user data. An email is sent when the archive is ready.

parameter user_id
User id

val delete_user_export : user_id:string -> id:string -> t -> unit -> Jsont.json

Delete a user export

**PeerTube >= 6.1**

parameter user_id
User id

parameter id
Entity id

val user_import_resumable_init : user_id:string -> body:UserImportResumable.T.t -> t -> unit -> Jsont.json

Initialize the resumable user import

**PeerTube >= 6.1** Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to initialize the import of the archive

parameter user_id
User id

val user_import_resumable : user_id:string -> upload_id:string -> t -> unit -> Jsont.json

Send chunk for the resumable user import

**PeerTube >= 6.1** Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to continue, pause or resume the import of the archive

parameter user_id
User id

parameter upload_id
Created session id to proceed with. If you didn't send chunks in the last hour, it is not valid anymore and you need to initialize a new upload.

val user_import_resumable_cancel : user_id:string -> upload_id:string -> t -> unit -> Jsont.json

Cancel the resumable user import

**PeerTube >= 6.1** Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to cancel the resumable user import

parameter user_id
User id

parameter upload_id
Created session id to proceed with. If you didn't send chunks in the last hour, it is not valid anymore and you need to initialize a new upload.

val get_latest_user_import : user_id:string -> t -> unit -> Jsont.json

Get latest user import

**PeerTube >= 6.1**

parameter user_id
User id

val add_video_channel_sync : body:VideoChannelSync.Create.t -> t -> unit -> Jsont.json

Create a synchronization for a video channel

val del_video_channel_sync : channel_sync_id:string -> t -> unit -> Jsont.json

Delete a video channel synchronization

parameter channel_sync_id
Channel Sync id

val trigger_video_channel_sync : channel_sync_id:string -> t -> unit -> Jsont.json

Triggers the channel synchronization job, fetching all the videos from the remote channel

parameter channel_sync_id
Channel Sync id

val add_video_channel : body:VideoChannel.Create.t -> t -> unit -> Jsont.json

Create a video channel

val put_video_channel : channel_handle:string -> body:VideoChannel.Update.t -> t -> unit -> Jsont.json

Update a video channel

parameter channel_handle
The video channel handle

val del_video_channel : channel_handle:string -> t -> unit -> Jsont.json

Delete a video channel

parameter channel_handle
The video channel handle

val delete_api_v1_video_channels_avatar : channel_handle:string -> t -> unit -> Jsont.json

Delete channel avatar

parameter channel_handle
The video channel handle

val post_api_v1_video_channels_avatar_pick : channel_handle:string -> t -> unit -> Jsont.json

Update channel avatar

parameter channel_handle
The video channel handle

val delete_api_v1_video_channels_banner : channel_handle:string -> t -> unit -> Jsont.json

Delete channel banner

parameter channel_handle
The video channel handle

val post_api_v1_video_channels_banner_pick : channel_handle:string -> t -> unit -> Jsont.json

Update channel banner

parameter channel_handle
The video channel handle

val list_video_channel_collaborators : channel_handle:string -> t -> unit -> Jsont.json

*List channel collaborators

**PeerTube >= 8.0**

parameter channel_handle
The video channel handle

val invite_video_channel_collaborator : channel_handle:string -> t -> unit -> Jsont.json

Invite a collaborator

**PeerTube >= 8.0** Invite a local user to collaborate on the specified video channel.

parameter channel_handle
The video channel handle

val remove_video_channel_collaborator : channel_handle:string -> collaborator_id:string -> t -> unit -> Jsont.json

Remove a channel collaborator

**PeerTube >= 8.0** Only the channel owner or the collaborator themselves can remove a collaborator from a channel

parameter channel_handle
The video channel handle

parameter collaborator_id
The collaborator id

val accept_video_channel_collaborator : channel_handle:string -> collaborator_id:string -> t -> unit -> Jsont.json

Accept a collaboration invitation

**PeerTube >= 8.0**

parameter channel_handle
The video channel handle

parameter collaborator_id
The collaborator id

val reject_video_channel_collaborator : channel_handle:string -> collaborator_id:string -> t -> unit -> Jsont.json

Reject a collaboration invitation

**PeerTube >= 8.0**

parameter channel_handle
The video channel handle

parameter collaborator_id
The collaborator id

val get_video_channel_followers : channel_handle:string -> ?start:string -> ?count:string -> ?sort:string -> ?search:string -> t -> unit -> Jsont.json

List followers of a video channel

parameter channel_handle
The video channel handle

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort followers by criteria

parameter search
Plain text search, applied to various parts of the model depending on endpoint

val post_api_v1_video_channels_import_videos : channel_handle:string -> body:ImportVideosInChannel.Create.t -> t -> unit -> Jsont.json

Import videos in channel

Import a remote channel/playlist videos into a channel

parameter channel_handle
The video channel handle

val get_api_v1_video_channels_video_playlists : channel_handle:string -> ?start:string -> ?count:string -> ?sort:string -> ?playlist_type:string -> t -> unit -> Jsont.json

List playlists of a channel

parameter channel_handle
The video channel handle

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val reorder_video_playlists_of_channel : channel_handle:string -> t -> unit -> Jsont.json

Reorder channel playlists

parameter channel_handle
The video channel handle

val get_playlists : ?start:string -> ?count:string -> ?sort:string -> ?playlist_type:string -> t -> unit -> Jsont.json

List video playlists

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val add_playlist : t -> unit -> Jsont.json

Create a video playlist

If the video playlist is set as public, `videoChannelId` is mandatory.

val get_playlist_privacy_policies : t -> unit -> Jsont.json

List available playlist privacy policies

val put_api_v1_video_playlists : playlist_id:string -> t -> unit -> Jsont.json

Update a video playlist

If the video playlist is set as public, the playlist must have a assigned channel.

parameter playlist_id
Playlist id

val delete_api_v1_video_playlists : playlist_id:string -> t -> unit -> Jsont.json

Delete a video playlist

parameter playlist_id
Playlist id

val get_video_playlist_videos : playlist_id:string -> ?start:string -> ?count:string -> t -> unit -> Jsont.json

List videos of a playlist

parameter playlist_id
Playlist id

parameter start
Offset used to paginate results

parameter count
Number of items to return

val add_video_playlist_video : playlist_id:string -> t -> unit -> Jsont.json

Add a video in a playlist

parameter playlist_id
Playlist id

val reorder_video_playlist : playlist_id:string -> t -> unit -> Jsont.json

Reorder playlist elements

parameter playlist_id
Playlist id

val put_video_playlist_video : playlist_id:string -> playlist_element_id:string -> t -> unit -> Jsont.json

Update a playlist element

parameter playlist_id
Playlist id

parameter playlist_element_id
Playlist element id

val del_video_playlist_video : playlist_id:string -> playlist_element_id:string -> t -> unit -> Jsont.json

Delete an element from a playlist

parameter playlist_id
Playlist id

parameter playlist_element_id
Playlist element id

val get_video_blocks : ?type_:string -> ?search:string -> ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List video blocks

parameter type_
list only blocks that match this type:

`1`: manual block

`2`: automatic block that needs review

parameter search
plain search that will match with video titles, and more

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort blocklists by criteria

val get_categories : t -> unit -> Jsont.json

List available video categories

val get_api_v1_videos_comments : ?search:string -> ?search_account:string -> ?search_video:string -> ?video_id:string -> ?video_channel_id:string -> ?auto_tag_one_of:string -> ?is_local:string -> ?on_local_video:string -> t -> unit -> Jsont.json

List instance comments

parameter search
Plain text search, applied to various parts of the model depending on endpoint

parameter search_account
Filter comments by searching on the account

parameter search_video
Filter comments by searching on the video

parameter video_id
Limit results on this specific video

parameter video_channel_id
Limit results on this specific video channel

parameter auto_tag_one_of
**PeerTube >= 6.2** filter on comments that contain one of these automatic tags

parameter is_local
**PeerTube >= 4.0** Display only local or remote objects

parameter on_local_video
Display only objects of local or remote videos

val delete_api_v1_videos_imports : id:string -> t -> unit -> Jsont.json

Delete video import

Delete ended video import

parameter id
Entity id

val post_api_v1_videos_imports_cancel : id:string -> t -> unit -> Jsont.json

Cancel video import

Cancel a pending video import

parameter id
Entity id

val post_api_v1_videos_imports_retry : id:string -> t -> unit -> Jsont.json

Retry video import

**PeerTube >= 8.0** Retry a pending video import

parameter id
Entity id

val get_languages : t -> unit -> Jsont.json

List available video languages

val get_licences : t -> unit -> Jsont.json

List available video licences

val update_live_id : id:string -> body:LiveVideo.Update.t -> t -> unit -> Jsont.json

Update information about a live

parameter id
The object id, uuid or short uuid

val get_api_v1_videos_live_sessions : id:string -> ?sort:string -> t -> unit -> Jsont.json

List live sessions

List all sessions created in a particular live

parameter id
The object id, uuid or short uuid

parameter sort
Sort column

val get_api_v1_videos_ownership : t -> unit -> Jsont.json

List video ownership changes

val post_api_v1_videos_ownership_accept : id:string -> t -> unit -> Jsont.json

Accept ownership change request

parameter id
The object id, uuid or short uuid

val post_api_v1_videos_ownership_refuse : id:string -> t -> unit -> Jsont.json

Refuse ownership change request

parameter id
The object id, uuid or short uuid

val get_video_privacy_policies : t -> unit -> Jsont.json

List available video privacy policies

val upload_resumable_init : body:VideoUploadRequestResumable.T.t -> t -> unit -> Jsont.json

Initialize the resumable upload of a video

Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to initialize the upload of a video

val upload_resumable_cancel : upload_id:string -> t -> unit -> Jsont.json

Cancel the resumable upload of a video, deleting any data uploaded so far

Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to cancel the upload of a video

parameter upload_id
Created session id to proceed with. If you didn't send chunks in the last hour, it is not valid anymore and you need to initialize a new upload.

val put_video : id:string -> t -> unit -> Jsont.json

Update a video

parameter id
The object id, uuid or short uuid

val del_video : id:string -> t -> unit -> Jsont.json

Delete a video

parameter id
The object id, uuid or short uuid

val add_video_block : id:string -> t -> unit -> Jsont.json

Block a video

parameter id
The object id, uuid or short uuid

val del_video_block : id:string -> t -> unit -> Jsont.json

Unblock a video by its id

parameter id
The object id, uuid or short uuid

val get_video_captions : id:string -> t -> unit -> Jsont.json

List captions of a video

parameter id
The object id, uuid or short uuid

val generate_video_caption : id:string -> t -> unit -> Jsont.json

Generate a video caption

**PeerTube >= 6.2** This feature has to be enabled by the administrator

parameter id
The object id, uuid or short uuid

val add_video_caption : id:string -> caption_language:string -> t -> unit -> Jsont.json

Add or replace a video caption

parameter id
The object id, uuid or short uuid

parameter caption_language
The caption language

val del_video_caption : id:string -> caption_language:string -> t -> unit -> Jsont.json

Delete a video caption

parameter id
The object id, uuid or short uuid

parameter caption_language
The caption language

val replace_video_chapters : id:string -> t -> unit -> Jsont.json

Replace video chapters

**PeerTube >= 6.0**

parameter id
The object id, uuid or short uuid

val delete_api_v1_videos_comments : id:string -> comment_id:string -> t -> unit -> Jsont.json

Delete a comment or a reply

parameter id
The object id, uuid or short uuid

parameter comment_id
The comment id

val post_api_v1_videos_comments_approve : id:string -> comment_id:string -> t -> unit -> Jsont.json

Approve a comment

**PeerTube >= 6.2** Approve a comment that requires a review

parameter id
The object id, uuid or short uuid

parameter comment_id
The comment id

val get_video_desc : id:string -> t -> unit -> Jsont.json

Get complete video description

parameter id
The object id, uuid or short uuid

val post_api_v1_videos_give_ownership : id:string -> t -> unit -> Jsont.json

Request ownership change

parameter id
The object id, uuid or short uuid

val del_video_hls : id:string -> t -> unit -> Jsont.json

Delete video HLS files

parameter id
The object id, uuid or short uuid

val list_video_passwords : id:string -> ?start:string -> ?count:string -> ?sort:string -> t -> unit -> Jsont.json

List video passwords

**PeerTube >= 6.0**

parameter id
The object id, uuid or short uuid

parameter start
Offset used to paginate results

parameter count
Number of items to return

parameter sort
Sort column

val add_video_password : id:string -> t -> unit -> Jsont.json

Add a video password

**PeerTube >= 8.0**

parameter id
The object id, uuid or short uuid

val update_video_password_list : id:string -> t -> unit -> Jsont.json

Update video passwords

**PeerTube >= 6.0**

parameter id
The object id, uuid or short uuid

val remove_video_password : id:string -> video_password_id:string -> t -> unit -> Jsont.json

Delete a video password

**PeerTube >= 6.0**

parameter id
The object id, uuid or short uuid

parameter video_password_id
The video password id

val put_api_v1_videos_rate : id:string -> t -> unit -> Jsont.json

Like/dislike a video

parameter id
The object id, uuid or short uuid

val delete_video_source_file : id:string -> t -> unit -> Jsont.json

Delete video source file

parameter id
The object id, uuid or short uuid

val replace_video_source_resumable_init : id:string -> body:VideoReplaceSourceRequestResumable.T.t -> t -> unit -> Jsont.json

Initialize the resumable replacement of a video

**PeerTube >= 6.0** Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to initialize the replacement of a video

parameter id
The object id, uuid or short uuid

val replace_video_source_resumable : id:string -> upload_id:string -> t -> unit -> Jsont.json

Send chunk for the resumable replacement of a video

**PeerTube >= 6.0** Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to continue, pause or resume the replacement of a video

parameter id
The object id, uuid or short uuid

parameter upload_id
Created session id to proceed with. If you didn't send chunks in the last hour, it is not valid anymore and you need to initialize a new upload.

val replace_video_source_resumable_cancel : id:string -> upload_id:string -> t -> unit -> Jsont.json

Cancel the resumable replacement of a video

**PeerTube >= 6.0** Uses a resumable protocol(https://github.com/kukhariev/node-uploadx/blob/master/proto.md) to cancel the replacement of a video

parameter id
The object id, uuid or short uuid

parameter upload_id
Created session id to proceed with. If you didn't send chunks in the last hour, it is not valid anymore and you need to initialize a new upload.

val list_video_storyboards : id:string -> t -> unit -> Jsont.json

List storyboards of a video

**PeerTube >= 6.0**

parameter id
The object id, uuid or short uuid

val post_api_v1_videos_studio_edit : id:string -> t -> unit -> Jsont.json

Create a studio task

Create a task to edit a video (cut, add intro/outro etc)

parameter id
The object id, uuid or short uuid

val create_video_transcoding : id:string -> t -> unit -> Jsont.json

Create a transcoding job

parameter id
The object id, uuid or short uuid

val add_view : id:string -> body:UserViewingVideo.T.t -> t -> unit -> Jsont.json

Notify user is watching a video

Call this endpoint regularly (every 5-10 seconds for example) to notify the server the user is watching the video. After a while, PeerTube will increase video's viewers counter. If the user is authenticated, PeerTube will also store the current player time.

parameter id
The object id, uuid or short uuid

val del_video_web_videos : id:string -> t -> unit -> Jsont.json

Delete video Web Video files

**PeerTube >= 6.0**

parameter id
The object id, uuid or short uuid

val get_api_v1_watched_words_accounts_lists : account_name:string -> t -> unit -> Jsont.json

List account watched words

**PeerTube >= 6.2**

parameter account_name
account name to list watched words

val post_api_v1_watched_words_accounts_lists : account_name:string -> t -> unit -> Jsont.json

Add account watched words

**PeerTube >= 6.2**

val put_api_v1_watched_words_accounts_lists : account_name:string -> list_id:string -> t -> unit -> Jsont.json

Update account watched words

**PeerTube >= 6.2**

parameter list_id
list of watched words to update

val delete_api_v1_watched_words_accounts_lists : account_name:string -> list_id:string -> t -> unit -> Jsont.json

Delete account watched words

**PeerTube >= 6.2**

parameter list_id
list of watched words to delete

val get_api_v1_watched_words_server_lists : t -> unit -> Jsont.json

List server watched words

**PeerTube >= 6.2**

val post_api_v1_watched_words_server_lists : t -> unit -> Jsont.json

Add server watched words

**PeerTube >= 6.2**

val put_api_v1_watched_words_server_lists : list_id:string -> t -> unit -> Jsont.json

Update server watched words

**PeerTube >= 6.2**

parameter list_id
list of watched words to update

val delete_api_v1_watched_words_server_lists : list_id:string -> t -> unit -> Jsont.json

Delete server watched words

**PeerTube >= 6.2**

parameter list_id
list of watched words to delete

val get_download_videos_generate : video_id:string -> video_file_ids:string -> ?video_file_token:string -> t -> unit -> Jsont.json

Download video file

Generate a mp4 container that contains at most 1 video stream and at most 1 audio stream. Mainly used to merge the HLS audio only video file and the HLS video only resolution file.

parameter video_id
The video id

parameter video_file_ids
streams of video files to mux in the output

parameter video_file_token
Video file token generated(#operation/requestVideoToken) by PeerTube so you don't need to provide an OAuth token in the request header.

val get_videos_podcast_feed : video_channel_id:string -> t -> unit -> Jsont.json

Videos podcast feed

parameter video_channel_id
Limit listing to a specific video channel

val get_syndicated_subscription_videos : format:string -> account_id:string -> token:string -> ?sort:string -> ?nsfw:string -> ?is_local:string -> ?include_:string -> ?privacy_one_of:string -> ?has_hlsfiles:string -> ?has_web_video_files:string -> t -> unit -> Jsont.json

Videos of subscriptions feeds

parameter format
format expected (we focus on making `rss` the most feature-rich ; it serves Media RSS(https://www.rssboard.org/media-rss))

parameter account_id
limit listing to a specific account

parameter token
private token allowing access

parameter sort
Sort column

parameter nsfw
whether to include nsfw videos, if any

parameter is_local
**PeerTube >= 4.0** Display only local or remote objects

parameter include_
**Only administrators and moderators can use this parameter**

Include additional videos in results (can be combined using bitwise or operator)

`0` NONE

`1` NOT_PUBLISHED_STATE

`2` BLACKLISTED

`4` BLOCKED_OWNER

`8` FILES

`16` CAPTIONS

`32` VIDEO SOURCE

parameter privacy_one_of
**PeerTube >= 4.0** Display only videos in this specific privacy/privacies

parameter has_hlsfiles
**PeerTube >= 4.0** Display only videos that have HLS files

parameter has_web_video_files
**PeerTube >= 6.0** Display only videos that have Web Video files

val get_syndicated_comments : format:string -> ?video_id:string -> ?account_id:string -> ?account_name:string -> ?video_channel_id:string -> ?video_channel_name:string -> t -> unit -> Jsont.json

Comments on videos feeds

parameter format
format expected (we focus on making `rss` the most feature-rich ; it serves Media RSS(https://www.rssboard.org/media-rss))

parameter video_id
limit listing comments to a specific video

parameter account_id
limit listing comments to videos of a specific account

parameter account_name
limit listing comments to videos of a specific account

parameter video_channel_id
limit listing comments to videos of a specific video channel

parameter video_channel_name
limit listing comments to videos of a specific video channel

val get_syndicated_videos : format:string -> ?account_id:string -> ?account_name:string -> ?video_channel_id:string -> ?video_channel_name:string -> ?sort:string -> ?nsfw:string -> ?is_local:string -> ?include_:string -> ?privacy_one_of:string -> ?has_hlsfiles:string -> ?has_web_video_files:string -> t -> unit -> Jsont.json

Common videos feeds

parameter format
format expected (we focus on making `rss` the most feature-rich ; it serves Media RSS(https://www.rssboard.org/media-rss))

parameter account_id
limit listing to a specific account

parameter account_name
limit listing to a specific account

parameter video_channel_id
limit listing to a specific video channel

parameter video_channel_name
limit listing to a specific video channel

parameter sort
Sort column

parameter nsfw
whether to include nsfw videos, if any

parameter is_local
**PeerTube >= 4.0** Display only local or remote objects

parameter include_
**Only administrators and moderators can use this parameter**

Include additional videos in results (can be combined using bitwise or operator)

`0` NONE

`1` NOT_PUBLISHED_STATE

`2` BLACKLISTED

`4` BLOCKED_OWNER

`8` FILES

`16` CAPTIONS

`32` VIDEO SOURCE

parameter privacy_one_of
**PeerTube >= 4.0** Display only videos in this specific privacy/privacies

parameter has_hlsfiles
**PeerTube >= 4.0** Display only videos that have HLS files

parameter has_web_video_files
**PeerTube >= 6.0** Display only videos that have Web Video files

val get_static_streaming_playlists_hls_private_ : filename:string -> ?video_file_token:string -> ?reinject_video_file_token:string -> t -> unit -> Jsont.json

Get private HLS video file

parameter filename
Filename

parameter video_file_token
Video file token generated(#operation/requestVideoToken) by PeerTube so you don't need to provide an OAuth token in the request header.

parameter reinject_video_file_token
Ask the server to reinject videoFileToken in URLs in m3u8 playlist

val get_static_streaming_playlists_hls : filename:string -> t -> unit -> Jsont.json

Get public HLS video file

parameter filename
Filename

val get_static_web_videos_private_ : filename:string -> ?video_file_token:string -> t -> unit -> Jsont.json

Get private Web Video file

**PeerTube >= 6.0**

parameter filename
Filename

parameter video_file_token
Video file token generated(#operation/requestVideoToken) by PeerTube so you don't need to provide an OAuth token in the request header.

val get_static_web_videos : filename:string -> t -> unit -> Jsont.json

Get public Web Video file

**PeerTube >= 6.0**

parameter filename
Filename